fix/ ajustes nos botões de todas as telas
