+++ frontend/src/components/Cadastro.jsx
@@ -1,47 +1,47 @@
-import { TextField, Button, Alert } from "@mui/material";
+import {TextField, Button, Alert} from "@mui/material";
 import axios from "axios";
-import { useState } from "react";
-import { useNavigate } from 'react-router-dom';
+import {useState} from "react";
+import {useNavigate} from 'react-router-dom';
 
 const Cadastro = () => {
     const navigate = useNavigate();
 
     const [formData, setFormData] = useState({
-        nome:'',
+        nome: '',
         login: '',
-        telefone:'',
+        telefone: '',
         senha: ''
     });
 
-    const [erroLogin, setErroLogin] = useState(false); 
+    const [erroLogin, setErroLogin] = useState(false);
 
     const handleChange = (e) => {
         setFormData({
-          ...formData,
-          [e.target.name]: e.target.value
+            ...formData,
+            [e.target.name]: e.target.value
         });
     };
 
-    
+
     const handleSubmit = (e) => {
         e.preventDefault();
-        
+
         axios.post('http://localhost:8080/usuarios/cadastro', formData, {
-          headers: {
-            'Content-Type': 'application/json'
-          }
-        })        
-        .then(res => {
-            setFormData({
-                nome:'',
-                login: '',
-                telefone:'',
-                senha: ''
-            });
-            navigate('/')
-            console.log(res);
+            headers: {
+                'Content-Type': 'application/json'
+            }
         })
-        
+            .then(res => {
+                setFormData({
+                    nome: '',
+                    login: '',
+                    telefone: '',
+                    senha: ''
+                });
+                navigate('/')
+                console.log(res);
+            })
+
 
     };
 
@@ -49,7 +49,7 @@ const Cadastro = () => {
         navigate('/');
     }
 
-    return(
+    return (
         <div style={{
             display: 'flex',
             justifyContent: 'center',
@@ -57,144 +57,144 @@ const Cadastro = () => {
             height: '100vh',
             width: '100vw',
             backgroundColor: '#f5f5f5'
-          }}>
+        }}>
             <div style={{
-              maxWidth: '420px',
-              margin: '40px auto',
-              padding: '30px',
-              backgroundColor: '#fff',
-              borderRadius: '20px',
-              boxShadow: '0 8px 24px rgba(0,0,0,0.1)',
-              fontFamily: 'roboto'
+                maxWidth: '420px',
+                margin: '40px auto',
+                padding: '30px',
+                backgroundColor: '#fff',
+                borderRadius: '20px',
+                boxShadow: '0 8px 24px rgba(0,0,0,0.1)',
+                fontFamily: 'roboto'
             }}>
-              <h2 style={{
-                textAlign: 'center',
-                fontSize: '24px',
-                fontWeight: '600',
-                color: '#000',
-                marginBottom: '20px'
-              }}>
-                Cadastrar-se
-              </h2>
-      
-              {erroLogin && (
-                <Alert severity="error" sx={{ mb: 2 }}>
-                  Numero já existe.
-                </Alert>
-              )}
-      
-              <form onSubmit={handleSubmit}>
-              <TextField
-                  id="nome"
-                  label="Nome"
-                  name="nome"
-                  value={formData.nome}
-                  onChange={handleChange}
-                  variant="outlined"
-                  fullWidth
-                  required
-                  sx={{
-                    marginBottom: '16px',
-                    '& .MuiOutlinedInput-root': {
-                      borderRadius: '12px'
-                    }
-                  }}
-                />
-                <TextField
-                  id="login"
-                  label="Login"
-                  name="login"
-                  value={formData.login}
-                  onChange={handleChange}
-                  variant="outlined"
-                  fullWidth
-                  required
-                  sx={{
-                    marginBottom: '16px',
-                    '& .MuiOutlinedInput-root': {
-                      borderRadius: '12px'
-                    }
-                  }}
-                />
-                <TextField
-                  id="telefone"
-                  label="Telefone"
-                  name="telefone"
-                  value={formData.telefone}
-                  onChange={handleChange}
-                  variant="outlined"
-                  fullWidth
-                  required
-                  sx={{
-                    marginBottom: '16px',
-                    '& .MuiOutlinedInput-root': {
-                      borderRadius: '12px'
-                    }
-                  }}
-                />
-                <TextField
-                  id="outlined-password-input"
-                  label="Senha"
-                  name="senha"
-                  type="password"
-                  autoComplete="current-password"
-                  value={formData.senha}
-                  onChange={handleChange}
-                  variant="outlined"
-                  fullWidth
-                  required
-                  sx={{
-                    marginBottom: '16px',
-                    '& .MuiOutlinedInput-root': {
-                      borderRadius: '12px'
-                    }
-                  }}
-                />
-      
-                <Button
-                  variant="contained"
-                  color="primary"
-                  type="submit"
-                  fullWidth
-                  sx={{
-                    backgroundColor: '#007AFF',
-                    borderRadius: '12px',
-                    textTransform: 'none',
+                <h2 style={{
+                    textAlign: 'center',
+                    fontSize: '24px',
                     fontWeight: '600',
-                    fontSize: '16px',
-                    padding: '10px 0',
-                    '&:hover': {
-                      backgroundColor: '#005FCC'
-                    }
-                  }}
-                >
-                  Cadastrar-se
+                    color: '#000',
+                    marginBottom: '20px'
+                }}>
+                    Cadastrar-se
+                </h2>
+
+                {erroLogin && (
+                    <Alert severity="error" sx={{mb: 2}}>
+                        Numero já existe.
+                    </Alert>
+                )}
+
+                <form onSubmit={handleSubmit}>
+                    <TextField
+                        id="nome"
+                        label="Nome"
+                        name="nome"
+                        value={formData.nome}
+                        onChange={handleChange}
+                        variant="outlined"
+                        fullWidth
+                        required
+                        sx={{
+                            marginBottom: '16px',
+                            '& .MuiOutlinedInput-root': {
+                                borderRadius: '12px'
+                            }
+                        }}
+                    />
+                    <TextField
+                        id="login"
+                        label="Login"
+                        name="login"
+                        value={formData.login}
+                        onChange={handleChange}
+                        variant="outlined"
+                        fullWidth
+                        required
+                        sx={{
+                            marginBottom: '16px',
+                            '& .MuiOutlinedInput-root': {
+                                borderRadius: '12px'
+                            }
+                        }}
+                    />
+                    <TextField
+                        id="telefone"
+                        label="Telefone"
+                        name="telefone"
+                        value={formData.telefone}
+                        onChange={handleChange}
+                        variant="outlined"
+                        fullWidth
+                        required
+                        sx={{
+                            marginBottom: '16px',
+                            '& .MuiOutlinedInput-root': {
+                                borderRadius: '12px'
+                            }
+                        }}
+                    />
+                    <TextField
+                        id="outlined-password-input"
+                        label="Senha"
+                        name="senha"
+                        type="password"
+                        autoComplete="current-password"
+                        value={formData.senha}
+                        onChange={handleChange}
+                        variant="outlined"
+                        fullWidth
+                        required
+                        sx={{
+                            marginBottom: '16px',
+                            '& .MuiOutlinedInput-root': {
+                                borderRadius: '12px'
+                            }
+                        }}
+                    />
+
+                    <Button
+                        variant="contained"
+                        color="primary"
+                        type="submit"
+                        fullWidth
+                        sx={{
+                            backgroundColor: '#007AFF',
+                            borderRadius: '12px',
+                            textTransform: 'none',
+                            fontWeight: '600',
+                            fontSize: '16px',
+                            padding: '10px 0',
+                            '&:hover': {
+                                backgroundColor: '#005FCC'
+                            }
+                        }}
+                    >
+                        Cadastrar-se
+                    </Button>
+                </form>
+                <Button
+                    variant="outlined"
+                    color="primary"
+                    type="submit"
+                    fullWidth
+
+                    sx={{
+                        borderRadius: '12px',
+                        textTransform: 'none',
+                        fontWeight: '600',
+                        fontSize: '16px',
+                        padding: '10px 0',
+                        marginTop: '10px',
+                        '&:hover': {
+                            backgroundColor: '#ccdff4',
+                            borderColor: '#005FCC'
+                        }
+                    }}
+
+                    onClick={redirectLogin}>
+                    Já tem uma conta? Login
                 </Button>
-              </form>
-              <Button 
-                variant="outlined"
-                color="primary"
-                type="submit"
-                fullWidth
-      
-                sx={{
-                  borderRadius: '12px',
-                  textTransform: 'none',
-                  fontWeight: '600',
-                  fontSize: '16px',
-                  padding: '10px 0',
-                  marginTop: '10px',
-                  '&:hover': {
-                      backgroundColor: '#ccdff4',
-                      borderColor: '#005FCC'
-                  }
-                  }}
-      
-                onClick={redirectLogin}> 
-                Já tem uma conta? Login
-              </Button>
             </div>
-          </div>
+        </div>
     );
 }
